woring on rest api

+++ src/main/java/de/jannik/hobbies/rest/ApiUser.java
@@ -1,24 +1,17 @@
 package de.jannik.hobbies.rest;
 
 import de.jannik.hobbies.model.entity.User;
-import de.jannik.hobbies.model.enums.Gender;
 import de.jannik.hobbies.service.UserService;
-import org.hibernate.id.insert.Binder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.FileNotFoundException;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.*;
 
 @RestController
 @RequestMapping("api/user")
-public class Api
+public class ApiUser
 {
   @Autowired
   UserService userService;
@@ -30,16 +23,27 @@ public class Api
     return ResponseEntity.status(HttpStatus.CREATED).body(save);
   }
   //toDo: warp this in a response entity
-  @GetMapping("/{id}")
-  public void delete(@PathVariable Long id)
+  @DeleteMapping("/{id}")
+  public ResponseEntity<Boolean> delete(@PathVariable Long id)
   {
+    if(userService.findById(id)== null)
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
     userService.deleteById(id);
+    return ResponseEntity.status(HttpStatus.OK).body(true);
   }
-  @PutMapping("/{id}")
-  public ResponseEntity<User> update(@PathVariable Long id)
+  @PatchMapping("/{id}")
+  public ResponseEntity<?> update(@RequestBody User partialUpdate ,@PathVariable Long id)
   {
 
     return null;
+  }
+  @GetMapping("/{id}")
+  public ResponseEntity<User> read(@PathVariable Long id)
+  {
+    User read = userService.findById(id);
+    if(read== null)
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+    return ResponseEntity.status(HttpStatus.OK).body(read);
   }
 
 }
